Move React Table fixture and helper method to fixtureHelpers file
